use actual epoch time

+++ NERODevelopment/src/controllers/buttoncontroller.h
@@ -23,12 +23,12 @@ public slots:
 protected:
   Model *m_model;
   int m_pageIndex;
-  long m_enterButtonDebounce = 0;
-  long m_upButtonDebounce = 0;
-  long m_downButtonDebounce = 0;
-  long m_rightButtonDebounce = 0;
-  long m_leftButtonDebounce = 0;
-  long m_homeButtonDebounce = 0;
+  int m_enterButtonDebounce = 0;
+  int m_upButtonDebounce = 0;
+  int m_downButtonDebounce = 0;
+  int m_rightButtonDebounce = 0;
+  int m_leftButtonDebounce = 0;
+  int m_homeButtonDebounce = 0;
   int m_debounceOffset = 100;
 };
